Fixing RegressionMeasure. Subgroup and complement now seem to be correctly computed. Not sure about all the statistics yet. Regression lines are printed to console.
git-svn-id: https://svn.science.uu.nl/repos/sci.knobb101.SubgroupDiscovery@807 3a85a776-1598-11dc-ad04-adf9e166afdd

+++ SubDisc/src/nl/liacs/subdisc/DataPoint.java
@@ -7,21 +7,21 @@ package nl.liacs.subdisc;
 
 public class DataPoint
 {
-	private final double itsX;
-	private final double itsY;
-	
-	public DataPoint(double theX, double theY)
+	private final float itsX;
+	private final float itsY;
+
+	public DataPoint(float theX, float theY)
 	{
 		itsX = theX;
 		itsY = theY;
 	}
 
-	public double getX()
+	public float getX()
 	{
-		return itsX;		
+		return itsX;
 	}
 
-	public double getY()
+	public float getY()
 	{
 		return itsY;
 	}
